backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.4
opencv-python-headless==4.11.0.86
numpy==2.2.5
pytesseract==0.3.13
PyMuPDF==1.24.11
supabase==2.15.2
python-dotenv==1.1.0
